Added makefile and several bug fixes detected at compilation time

+++ src/dynamic_sh_mem.h
@@ -11,6 +11,6 @@
 #define DYNAMIC_SH_MEM_H
 
 // variable for allowing dynamic allocation of __shared__ memory (used in charge_deposition, poisson_solver, )
-extern __shared__ double sh_mem[];
+extern __shared__ float sh_mem[];
 
 #endif
